pyproject.toml
[tool.poetry]
name = "python-voice-assistant"
version = "0.1.0"
description = ""
license = "GPL-3.0-only"
authors = ["YisusChrist <alejandro.gonzalezm02@estudiante.uam.es>"]
readme = "README.md"
repository = "https://github.com/yisuschrist/python-voice-assistant"

[tool.poetry.dependencies]
python = "^3.12"
newsapi-python = "^0.2.7"
psutil = "^5.9.8"
pyaudio = "^0.2.14"
python-dotenv = "^1.0.1"
pyttsx3 = "^2.90"
setuptools = "^69.1.0"
speechrecognition = "^3.10.1"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
python-voice-assistant = "python_voice_assistant.__main__:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
